fix completely broken logic of processing pending outbound packets.  We now do all the req_ctx state changes from within pcd_handler.c and only expose the refill_ep() functions.
It's questionable whether we need to keep calling those functions all the time
either.  Once kicked off, the tx completion interrupts of the UDC will refill
the pipe as long as there is data available.

So realistically speaking, we'd only need to trigger tx once (and keep
triggering until no EBUSY or other error is signalled anymore).  Some form of
state machine would be fine for doing this, one that registers a timer with
the core OS timer (once we have something equivalent to 'struct timer_list'.



git-svn-id: https://svn.openpcd.org:2342/trunk@217 6dc7ffe9-61d6-0310-9af1-9938baff3ed1