simtrace: implement SIM power switching API
there is now a high-level command by which we can swithc the VCC
supply of the SIM cards.

For some reason, the _PASS variant (passing through the voltage from the
phone) doesn't seem to work reliably.  It might be that we are draining
reverse current throught the LDO once we supply SIM power that way.