eabi: Remove EABI functions to fix building with EABI toolchains
These symbols are not referenced on an arm-elf gcc 4.7, they are
used when building with an arm-eabi gcc 4.7 and then cause linking
issues. The linking to libgcc can not be omitted as the irq code is
using ffs. Remove the EABI symbols. The alternative would be to
declare the EABI as weak inside the lib1funcs.

Link error:
multiple definition of `__udivsi3'
multiple definition of `__aeabi_uidiv'

Exported symbols:
nm lib/lib1funcs.o
00000000 t Ldiv0
00000010 T __div0