Ensure to transmit current req_ctx on RESET/VCC events
When any of the following events occur:
* VCC_PHONE off
* nRST
* RST

we tranmsit the current req_ctx, if there is data pending.

This ensures that for any successive data, the ATR flag in the next
req_ctx containing the data, not in the previous req_ctx.  It also
ensures that the ATR is aligned at 0 offset in the new req_ctx, which is
an assumption the host software makes but the previous code didn't
ensure.

Furthermore, we introduce a periodic flushing of any pending but
incomplete req_ctx.